+++ python/pfnet/cbus.pyx
@@ -467,24 +467,6 @@ cdef class Bus:
             else:
                 return np.array(r)
 
-    property index_vl:
-        """ Index of voltage low limit violation variable (int or array). """
-        def __get__(self):
-            r = [cbus.BUS_get_index_vl(self._c_ptr,t) for t in range(self.num_periods)]
-            if self.num_periods == 1:
-                return AttributeInt(r[0])
-            else:
-                return np.array(r)
-
-    property index_vh:
-        """ Index of voltage high limit violation variable (int or array). """
-        def __get__(self):
-            r = [cbus.BUS_get_index_vh(self._c_ptr,t) for t in range(self.num_periods)]
-            if self.num_periods == 1:
-                return AttributeInt(r[0])
-            else:
-                return np.array(r)
-
     property index_P:
         """ Index of bus active power mismatch (int). """
         def __get__(self): return cbus.BUS_get_index_P(self._c_ptr)